Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.7.1
PyJWT[crypto]==2.10.1
httpx==0.27.2
qdrant-client==1.12.1
sentence-transformers==3.3.1
pypdf==5.1.0
openai==1.59.7
huggingface-hub==0.27.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
